requirements.txt
streamlit
ultralytics
google-generativeai
ephem
grad-cam
fpdf
opencv-python-headless
Pillow
numpy